Adopt GetStarted component on all pages to require friend to split with
