+++ expense-splitter/src/components/friend/Friend.jsx
@@ -5,6 +5,7 @@ import SearchBar from "../ui/SearchBar";
 import Button from "../ui/Button";
 import FriendList from "./FriendList";
 import ButtonFooter from "../ui/ButtonFooter";
+import GetStarted from "../widgets/GetStarted";
 
 const Friend = () => {
   const navigate = useNavigate();
@@ -13,8 +14,8 @@ const Friend = () => {
   const [inputText, setInputText] = useState("");
 
   let inputHandler = (e) => {
-      var lowerCase = e.target.value.toLowerCase();
-      setInputText(lowerCase);
+    var lowerCase = e.target.value.toLowerCase();
+    setInputText(lowerCase);
   };
 
   useEffect(() => {
@@ -29,9 +30,9 @@ const Friend = () => {
       <>
         <h1 className="text-center">Friends</h1>
         <div className="mb-2">
-          <SearchBar input={inputText} inputHandler={inputHandler}/>
+          <SearchBar input={inputText} inputHandler={inputHandler} />
         </div>
-        <FriendList input={inputText}/>
+        <FriendList input={inputText} />
         <ButtonFooter>
           <Button
             className="bg-primary"
@@ -42,6 +43,7 @@ const Friend = () => {
             Add Friend
           </Button>
         </ButtonFooter>
+        <GetStarted />
       </>
     )
   );
